+++ backend/module-application/toonpick-app-api/src/main/java/com/toonpick/auth/service/UsernameLoginService.java
@@ -8,7 +8,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import com.toonpick.user.CustomUserDetails;
+import com.toonpick.dto.CustomUserDetails;
 
 @Service
 @RequiredArgsConstructor
