+++ backend/module-application/toonpick-app-api/src/main/java/com/toonpick/review/controller/AuthenticatedWebtoonReviewController.java
@@ -4,7 +4,7 @@ import com.toonpick.annotation.CurrentUser;
 import com.toonpick.review.request.WebtoonReviewCreateRequest;
 import com.toonpick.review.request.WebtoonReviewUpdateRequest;
 import com.toonpick.review.response.WebtoonReviewResponse;
-import com.toonpick.user.CustomUserDetails;
+import com.toonpick.dto.CustomUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
