+++ backend/module-application/toonpick-app-api/src/main/java/com/toonpick/review/controller/PublicWebtoonReviewController.java
@@ -3,7 +3,7 @@ package com.toonpick.review.controller;
 import com.toonpick.annotation.CurrentUser;
 import com.toonpick.dto.PagedResponseDTO;
 import com.toonpick.review.response.WebtoonReviewResponse;
-import com.toonpick.user.CustomUserDetails;
+import com.toonpick.dto.CustomUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
