+++ backend/module-internal/core-security/src/main/java/com/toonpick/dto/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package com.toonpick.user;
+package com.toonpick.dto;
 
 import lombok.Builder;
 import org.springframework.security.core.GrantedAuthority;
